improve buildscripts
