+++ .gitignore
@@ -287,4 +287,3 @@ Temporary Items
 .idea/
 run/
 target/
-gradle.properties
