+++ build-logic/src/main/kotlin/ignite.base-conventions.gradle.kts
@@ -1,0 +1,20 @@
+plugins {
+  id("net.kyori.indra")
+  id("net.kyori.indra.license-header")
+}
+
+repositories {
+  mavenLocal()
+  mavenCentral()
+  maven {
+    url = uri("https://oss.sonatype.org/content/groups/public/")
+  }
+}
+
+indra {
+  github("vectrix-space", "ignite") {
+    ci(true)
+  }
+
+  mitLicense()
+}
