+++ build-logic/src/main/kotlin/ignite.parent-build-logic.gradle.kts
@@ -1,0 +1,3 @@
+plugins {
+  id("net.kyori.indra.git")
+}
