+++ gradle.properties
@@ -1,0 +1,4 @@
+# Gradle Properties
+org.gradle.daemon=true
+org.gradle.configureondemand=true
+org.gradle.parallel=true
